fix(Programme/Shortcodes)!: Fixed an issue with Programme pages
Current implementations of programme shortcodes _might_ break since we have
changed how we output the pages, from a normal include, to use `ob_start()`
to collect the entire page, and then get the data with `ob_get_clean()`
and return the value.

BREAKING CHANGE: Custom code with `do_shortcode` now needs to do
`echo do_shortcode( '[eduadmin-programme-list]' );` after the change.
Which is the intended way to use `do_shortcode` the previous versions
were not working in the correct way.